3D Vis + Accuracy Increases for PhotonVision
Full 3D Visualization working with the PhotonVision Sim. I also noticed an error in the code from last year that set the detection mode to best instead of multi-tag solve pnp so we will see increased accuracy now. I also added the PhotonPoseEstimator to the pipeline instead of directly fusing with the WPIlib Kalman filter which is what is recommended in the photonVision docs and will also increase the accuracy as well.